Update DIG528-2 UDEV rules to use new naming scheme
This splits up the platform specific rules files into two, one for
creating symlinks and one for running scripts (and notifying the mote
host). The main reason is that they need to be run in sequence and the
priority system allows this to happen based on the naming of the rules
files.

Since no specification exists for how UDEV rules should be prioritized,
Re-Mote will adopt the one used by Debian. Add a small README file
documenting the new naming scheme and motivation.