Issue 10: Drop usage of tosaddress table in favor mote attributes usage
This changes Host::findOrCreateMote() to not create a new entry in the
tosaddress table just so it can assign an automatically incremented TOS
address to the mote. Similarly, known motes will be assigned a TOS
address based on the value in tosaddress mote attribute.

The main advantage is the simplified code because TOS address mote
attribute handling is now handled in one place and also only stored one
place in the databse. Furthermore, it will allow the tosaddress table to
be removed from remote-db. A drawback is that TOS addresses are no
longer stored in an integer representation, which puts more
responsability on administrators when changing the TOS address.