SerialControl: Improve error handling of system calls
This ensures that no pipe ends are leaked if fork fails, and that
dup2() problems will make the child exit immediately.

It also removes the error message printed to stderr in the child,
since the client already handles and notifies the user about this
and because any dup2() problems will cause that the messages will
never be printed anyway.