Consolidate mote initialization and validation
This changes the mote constructor to use Mote::validate() in order to
consolidate how mote creation and updating in terms of validations takes
place.

The path hugely simplifies the TTY update code by removing code that
checked for whether the path to the TTY had changed and in turn reopened
the TTY if necessarry. Since UDEV was introduced the TTY path is stable
and so this code has been obsolete for some time.

Instead, the UDEV TTY symlink target is resolved and used as the path to
the TTY and used for reenabling detection of whether the TTY has
changed.