Remove code to reopen renamed TTYs when validating motes
It is highly unlikely that a renamed or rather "relinked" TTY will have
gone undetected by select() and the following readBuf(). In which case
the TTY has already been closed.

The code most of all looks like a graceful handling of the above case
from when the TTY name/number was acquired from /proc and thus harder to
get.