Move stop() call from closeTty() to the only calling site needing it
The basic idea is to make it clearer what places calls stop and reduce
the number of places that calls stop() since clearly it is not always
desirable when calling closeTty. The only remaining real user of stop()
is MoteHost::program() to keep the mote in stop state when programming.

The places that does not need it is:

 - openTty() which would call closeTty() if its own stop() call failed.
 - readBuf() which closes the TTY port if read fails potentially leading
   to ioctl() failure.
 - ~SerialControl which closes the TTY port when the mote is destructed,
   which should only ever happen if the mote is unplugged or otherwise
   end up in an invalid state.