Postpone cleaning up of child until getChildResult() is called
This basically just moves the cleanUpProgram() a few cycles later with
respect to MoteHote::handleMoteData. Before it happened in readBuf() now
it happens a few lines later in handleMoteData when it calls
getChildResult(). Allows the childResult member to be removed.

It also feels better to postpone the waitpid() so that the SerialControl
destructor can safely assumes that cancelProgram() will not hang.