Consolidate mote programming in the Mote class
This somewhat lowlevel stuff really belongs in the Mote class and it
needs to be there to fix some of the requirements related to issue 6.
Additionally, it allows to the getProgrammerPath() accessor to be
removed use the more appropriate hasChild() instead of getStatus() to
check for an existing programming session.

To further improve robustness of programming the path of the programmer
is now also checked for existence by Mote::validate().